+++ src/pages/delivery/components/Map.tsx
@@ -1,0 +1,9 @@
+import { Box } from "@mui/material";
+
+export default function Map() {
+  return (
+    <Box position="absolute" zIndex="-1" width="100%" height="100%">
+        <Box component="img" width="100%" height="100%" src="https://s8.uupload.ir/files/group_3159_wiw5.png" />
+    </Box>
+  )
+}
